test(core): share the racing-double harness instead of copying it
The dups: similarity check flagged one clone introduced by this PR - 19 lines shared between
RetryQueueRequeueWindowTest and WorkManagerStaleCheckDoubleLookupTest. Real, and exactly the case
the report exists for: the new test copied the existing racing-double harness rather than extending
it, which is the "reuse test utilities - search before you add" rule going the wrong way.

RacingSeamWorkManager now holds the part that was duplicated - the one-shot arming mechanism (arm,
raceHasFired, fireOnceIfArmed) and the constructor. Each subclass keeps only what actually differs,
which is its seam: the pre-existing double overrides checkpoint 3's two-argument staleness lookup,
the new one overrides the live single-argument check and adds its own live-check bookkeeping.

This edits WorkManagerStaleCheckDoubleLookupTest, a pre-existing master file this branch had not
otherwise touched. That is the right place for the fix rather than a second copy, but it is worth
knowing when reading the diff. The change is mechanical: no assertion, arm or scenario altered on
either side, and the reasoning for keeping the fired flag separate from the armed slot moved onto
the base class rather than being deleted - a cleared armed slot cannot tell "armed, then fired" from
"never armed", so a precondition built on it would pass on a test that forgot to arm.

Not deduplicated, deliberately: the module-construction block (PCModuleTestEnv + MockConsumer +
PARTITION ordering) is pre-existing across nine test files on master, so it is base-branch
duplication and out of this PR's scope.

Both classes green with the same counts as before (8 and 5). Core unit suite 688 run / 0 failures /
8 skipped - unchanged, so nothing was lost in the move. Whole-reactor unit build green;
bin/check-all.sh 16 passed, 0 failed.
